// frontend/src/app/core/generated/circabc/model/applicant.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';


/**
 * an application for membership representation
 */
export interface Applicant { 
    submitted?: string;
    user?: User;
    justification?: string;
}