// frontend/src/app/core/generated/circabc/model/userProfile.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { Profile } from './profile';


/**
 * the representation of the membership of a user inside an Interest group
 */
export interface UserProfile { 
    user?: User;
    profile?: Profile;
    expirationDate?: string;
}